Extract HabanaModelAdapter from run_lm_eval to new script file.
To separate dependencies and make managing imports order easier,
HabanaModelAdapter is extracted and moved to model_adapter.py.
This allows us to cleanly initialize model and environmental
variables before importing lm_eval, transformers and torch,
all of which in turn load habana shared object, which needs to
happen after env vars are prepared.
